Refactor policies to need explicit specificatino of value function.
Simplify the inheritance structure for policies to have a single
virtual value function that is provided as a direct base at the
top of the value inheritance chain. So no more creating inplace
value functions.

Users now need to explicitly provide a correct Finite Value Fn
with a FiniteValue value type for creation to work. Helper methods
which speed this up have been created as well.
The change is made in order to enable complex combinations to
be direct base value functions. For example providing a combination
as a base instead of expecting something else.